+++ src/com/genir/aitweaks/core/state/Config.kt
@@ -20,6 +20,7 @@ class Config {
     val omniShieldKeybind: Int = LunaSettings.getInt("aitweaks", "aitweaks_omni_shield_keybind")!!
     val aimAssistKeybind: Int = LunaSettings.getInt("aitweaks", "aitweaks_aim_bot_keybind")!!
     val aimAssistRotateShip: Boolean = LunaSettings.getBoolean("aitweaks", "aitweaks_aim_bot_rotate_ship")!!
+    val aimAssistTargetJunk: Boolean = LunaSettings.getBoolean("aitweaks", "aitweaks_aim_bot_target_junk")!!
 
     // Debug
     val devMode: Boolean = LunaSettings.getBoolean("aitweaks", "aitweaks_enable_devmode")!!