+++ changelog.md
@@ -1,7 +1,8 @@
 - Custom AI: If the enemy is in full retreat, ships will attempt to intercept and block its escape route. Suggested by Reshy.
 - Custom AI: Several improvements and regression fixes in ship movement coordination.
 - All ammo-based kinetic weapons attack enemy ships only when their shields are raised. Previously the behavior applied only to vanilla Needlers. This feature can be disabled in LunaLib settings. Suggested by vinh.
-- Aim Assist can be configured not to target hulks and asteroids. Suggested by Roxorium.  
+- Aim Assist can be configured not to target hulks and asteroids. Suggested by Roxorium.
+- Potentially fixed an incompatibility with Luddic Enhancement: IED Edition, where "the funny Luddic explosive Tankers loses its suicidal behavior". Reported by TheShear.
 
 Version 1.12.1